Move things to a more sensible location
